# Conflicts:
#   backend/TmiProject/src/main/java/com/tmi/controller/app/AppController.java
#   backend/TmiProject/src/main/java/com/tmi/repository/AppRepository.java
#   frontend/src/pages/project/Project.jsx
#   frontend/src/utils/axios.js

+++ backend/TMI/src/main/java/com/tmi/tmi/model/Test.java
@@ -28,6 +28,6 @@ public class Test {
     int failCount;
     int errorCount;
     int skipCount;
-    float elapsedTime;
+    int elapsedTime;
     List<TestCase> testCaseList;
 }
